+++ src/Transformers/AssetsTransformer.php
@@ -2,8 +2,13 @@
 
 namespace Statamic\Importer\Transformers;
 
+use Illuminate\Support\Carbon;
 use Illuminate\Support\Facades\Http;
+use Statamic\Assets\AssetUploader;
+use Statamic\Contracts\Assets\AssetContainer as AssetContainerContract;
+use Statamic\Facades\Asset;
 use Statamic\Facades\AssetContainer;
+use Statamic\Facades\Path;
 use Statamic\Support\Str;
 
 class AssetsTransformer extends AbstractTransformer
@@ -16,6 +21,11 @@ class AssetsTransformer extends AbstractTransformer
 
         $baseUrl = $this->config('base_url');
         $relatedField = $this->config('related_field', 'path');
+
+        // When $value is a JSON string, decode it.
+        if (Str::startsWith($value, ['{', '[']) || Str::startsWith($value, ['[', ']'])) {
+            $value = collect(json_decode($value, true))->join('|');
+        }
 
         $assets = collect(explode('|', $value))->map(function ($path) use ($assetContainer, $relatedField, $baseUrl) {
             $path = Str::of($path)
@@ -34,8 +44,13 @@ class AssetsTransformer extends AbstractTransformer
                     return null;
                 }
 
-                $assetContainer->disk()->put($path, $request->body());
-                $asset = tap($assetContainer->makeAsset($path))->save();
+                $asset = Asset::make()
+                    ->container($assetContainer)
+                    ->path($this->assetPath($assetContainer, $path));
+
+                $assetContainer->disk()->put($asset->path(), $request->body());
+
+                $asset->save();
             }
 
             return $asset?->path();
@@ -44,13 +59,38 @@ class AssetsTransformer extends AbstractTransformer
         return $this->field->get('max_files') === 1 ? $assets->first() : $assets->all();
     }
 
+    private function assetPath(AssetContainerContract $assetContainer, string $path): string
+    {
+        $extension = pathinfo($path, PATHINFO_EXTENSION);
+
+        if (config('statamic.assets.lowercase')) {
+            $ext = strtolower($extension);
+        }
+
+        $filename = AssetUploader::getSafeFilename(pathinfo($path, PATHINFO_FILENAME));
+
+        $directory = $this->config('folder') ?? Str::beforeLast($path, '/');
+        $directory = ($directory === '.') ? '/' : $directory;
+
+        $path = Path::tidy($directory.'/'.$filename.'.'.$ext);
+        $path = ltrim($path, '/');
+
+        // If the file exists, we'll append a timestamp to prevent overwriting.
+        if ($assetContainer->disk()->exists($path)) {
+            $basename = $filename.'-'.Carbon::now()->timestamp.'.'.$ext;
+            $path = Str::removeLeft(Path::assemble($directory, $basename), '/');
+        }
+
+        return $path;
+    }
+
     public function fieldItems(): array
     {
         return [
             'related_field' => [
                 'type' => 'select',
                 'display' => __('Related Field'),
-                'instructions' => __('Which field does the data reference?'),
+                'instructions' => __('importer::messages.assets_related_field_instructions'),
                 'default' => 'url',
                 'options' => [
                     ['key' => 'path', 'value' => __('Path')],
@@ -61,14 +101,22 @@ class AssetsTransformer extends AbstractTransformer
             'base_url' => [
                 'type' => 'text',
                 'display' => __('Base URL'),
-                'instructions' => __('The base URL to prepend to the path.'),
+                'instructions' => __('importer::messages.assets_base_url_instructions'),
                 'if' => ['related_field' => 'url'],
             ],
             'download_when_missing' => [
                 'type' => 'toggle',
                 'display' => __('Download when missing?'),
-                'instructions' => __("If the asset can't be found in the asset container, should it be downloaded?"),
+                'instructions' => __('importer::messages.assets_download_when_missing_instructions'),
                 'if' => ['related_field' => 'url'],
+            ],
+            'folder' => [
+                'type' => 'asset_folder',
+                'display' => __('Folder'),
+                'instructions' => __('importer::messages.assets_folder_instructions'),
+                'if' => ['download_when_missing' => true],
+                'container' => $this->field->get('container'),
+                'max_items' => 1,
             ],
         ];
     }