+++ src/Transformers/EntriesTransformer.php
@@ -17,6 +17,11 @@ class EntriesTransformer extends AbstractTransformer
             $value = collect(unserialize($value))->join('|');
         }
 
+        // When $value is a JSON string, decode it.
+        if (Str::startsWith($value, ['{', '[']) || Str::startsWith($value, ['[', ']'])) {
+            $value = collect(json_decode($value, true))->join('|');
+        }
+
         if ($this->config('related_field') === 'id') {
             return is_string($value) ? explode('|', $value) : $value;
         }
@@ -75,7 +80,7 @@ class EntriesTransformer extends AbstractTransformer
             'related_field' => [
                 'type' => 'select',
                 'display' => __('Related Field'),
-                'instructions' => __('Which field does the data reference?'),
+                'instructions' => __('importer::messages.entries_related_field_instructions'),
                 'default' => 'id',
                 'options' => $fields
                     ->map(fn ($field) => ['key' => $field->handle(), 'value' => $field->display()])
@@ -87,7 +92,7 @@ class EntriesTransformer extends AbstractTransformer
             'create_when_missing' => [
                 'type' => 'toggle',
                 'display' => __('Create entry when missing?'),
-                'instructions' => __("Create the entry if it doesn't exist."),
+                'instructions' => __('importer::messages.entries_create_when_missing_instructions'),
                 'default' => false,
             ],
         ];