+++ tests/Transformers/DateTransformerTest.php
@@ -17,7 +17,7 @@ class DateTransformerTest extends TestCase
     public $blueprint;
     public $import;
 
-    public function setUp(): void
+    protected function setUp(): void
     {
         parent::setUp();